requirements.txt
# 自动化测试框架依赖包

# 核心测试框架
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist>=3.3.1
pytest-rerunfailures==12.0

# Allure报告
allure-pytest==2.13.2

# Web自动化
playwright==1.40.0

# HTTP请求
requests==2.31.0

# 数据处理
pyyaml==6.0.1
jsonpath-ng==1.6.0
openpyxl==3.1.2
pandas>=2.0.3

# 数据库连接
pymysql==1.1.0
psycopg2-binary==2.9.9

# 加密解密
cryptography>=41.0.3
pyjwt>=2.8.0

# 假数据生成
faker==19.12.0

# 日志处理
loguru==0.7.2

# 模板引擎
jinja2==3.1.2

# 系统监控
psutil>=5.9.5

# 图像处理
pillow>=10.0.0

# 类型提示
typing-extensions>=4.7.1

# 异步支持
aiohttp>=3.8.5

# 性能分析（可选）
memory-profiler>=0.61.0

# 代码质量检查（可选）
flake8>=6.0.0
black>=23.7.0

# 开发工具（可选）
ipython>=8.14.0

# 进度条和美化输出（可选）
tqdm>=4.66.1
rich>=13.5.2
colorama>=0.4.6

# 文件监控（配置热重载）
watchdog>=3.0.0

# 重试机制（可选）
tenacity>=8.2.3

# 调试工具（可选）
ipdb>=0.13.13